Create pizza directory (also rename modules)
