+++ .gitignore
@@ -24,4 +24,4 @@ go.work.sum
 # env file
 .env
 
-test_directory/
+test-directory/
