+++ mraziz/file.go
@@ -1,0 +1,6 @@
+package mraziz
+
+type File struct {
+	NameT        string
+	ContentTFile string
+}
