+++ test.sh
@@ -1,0 +1,14 @@
+#!/bin/bash
+
+go install .
+
+rm -rf test-directory
+mkdir test-directory
+
+
+pushd test-directory
+mraziz init
+popd
+
+
+stat test-directory/.pizza
